requirements.txt
Django==1.8.3
psycopg2==2.6.1
metron==1.3.5
django-jsonfield==0.9.13
djangorestframework==3.1.3
markdown==2.6.2
django-filter==0.10.0
django-allauth==0.21.0
django-rest-auth==0.4.0
django-avatar==2.0
django-photologue==3.3.2
django-extra-fields==0.3
django-oauth-toolkit==0.9.0
django_post_office==2.0.5
django-analytical==2.0.0
django-countries==3.4.1
django-widget-tweaks==1.4.1
django-cities-light==3.2.0 
django-attachments==1.0
stripe==1.34.0
django-zebra==0.4.5
pytz==2016.4
django-twilio==0.8.0
